+++ Common/Extensions.cs
@@ -6,17 +6,17 @@ namespace Alyn.Pointer.Common
 {
     public static class Extensions
     {
-        public static IEnumerable<Control> DescendentsFromPoint(this Control control, Point pt)
+        public static IEnumerable<Control> DescendentsFromPoint(this Control control, Point point)
         {
-            var ctrl = control.GetChildAtPoint(pt);
+            var ctrl = control.GetChildAtPoint(point);
             if (ctrl != null)
             {
                 yield return ctrl;
 
-                var screenPt = control.PointToScreen(pt);
-                pt = ctrl.PointToScreen(screenPt);
+                var screenPt = control.PointToScreen(point);
+                point = ctrl.PointToScreen(screenPt);
 
-                foreach (var item in DescendentsFromPoint(ctrl, pt))
+                foreach (var item in DescendentsFromPoint(ctrl, point))
                 {
                     yield return item;
                 }
