+++ Player/Alyn.Pointer.App.csproj
@@ -142,7 +142,6 @@
     <Compile Include="MainForm.Designer.cs">
       <DependentUpon>MainForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="OS.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <EmbeddedResource Include="MainForm.resx">
@@ -156,9 +155,6 @@
     </EmbeddedResource>
     <EmbeddedResource Include="SettingsForm.resx">
       <DependentUpon>SettingsForm.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="URLForm.resx">
-      <DependentUpon>URLForm.cs</DependentUpon>
     </EmbeddedResource>
     <Compile Include="Properties\Resources.Designer.cs">
       <AutoGen>True</AutoGen>
@@ -181,12 +177,6 @@
     </Compile>
     <Compile Include="SettingsForm.Designer.cs">
       <DependentUpon>SettingsForm.cs</DependentUpon>
-    </Compile>
-    <Compile Include="URLForm.cs">
-      <SubType>Form</SubType>
-    </Compile>
-    <Compile Include="URLForm.Designer.cs">
-      <DependentUpon>URLForm.cs</DependentUpon>
     </Compile>
   </ItemGroup>
   <ItemGroup>
