+++ Player/app.config
@@ -2,18 +2,18 @@
 <configuration>
   <configSections>
     <sectionGroup name="userSettings" type="System.Configuration.UserSettingsGroup, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" >
-      <section name="Player.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
+      <section name="Alyn.Pointer.App.Properties.Settings" type="System.Configuration.ClientSettingsSection, System, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" allowExeDefinition="MachineToLocalUser" requirePermission="false" />
     </sectionGroup>
   </configSections>
   <userSettings>
-    <Player.Properties.Settings>
+    <Alyn.Pointer.App.Properties.Settings>
       <setting name="SoundAlerts" serializeAs="String">
         <value>True</value>
       </setting>
       <setting name="gazeDelay" serializeAs="String">
         <value>2</value>
       </setting>
-    </Player.Properties.Settings>
+    </Alyn.Pointer.App.Properties.Settings>
   </userSettings>
   <startup>
     <supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.7.2"/>
